cambio tipo de variable de rubro insumo

+++ back/src/main/java/ebs/back/entity/RubroInsumo.java
@@ -17,7 +17,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 @Entity
 public class RubroInsumo implements Serializable {
 
-	private Float id;
+	private Long id;
 	private String denominacion;
 	private List<Insumo> insumos = new ArrayList<>();
 
@@ -25,7 +25,7 @@ public class RubroInsumo implements Serializable {
 
 	}
 
-	public RubroInsumo(Float id, String denominacion, List<Insumo> insumos) {
+	public RubroInsumo(Long id, String denominacion, List<Insumo> insumos) {
 		this.id = id;
 		this.denominacion = denominacion;
 		this.insumos = insumos;
@@ -35,11 +35,11 @@ public class RubroInsumo implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "idRubroInsumo")
-	public Float getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Float id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
